feature/(new-integrations): restructing freshmarketer code
